feat: 1411 clean up the webpack config for the project
