+++ package.json
@@ -167,9 +167,6 @@
     "node-sass": "^4.12.0",
     "prettier": "^1.19.1",
     "prettier-plugin-packagejson": "^2.0.1",
-    "purgecss": "^1.3.0",
-    "purgecss-whitelister": "^2.4.0",
-    "purify-css": "^1.2.6",
     "query-string": "5.0.0",
     "react-docgen-typescript-loader": "^3.3.0",
     "regenerator-runtime": "^0.13.5",
