+++ packages/site-builder/package.json
@@ -11,10 +11,10 @@
   "author": "Reapit Ltd",
   "main": "./src/index.ts",
   "scripts": {
-    "build:prod": "node '../../scripts/webpack/webpack-sass-prod/index.js'",
+    "build:prod": "node '../../scripts/webpack/webpack.config.prod.js'",
     "fetch-config": "yarn config-manager fetchConfig site-builder",
     "release:dev": "node ../../scripts/release/release-dev.js site-builder reapit-site-builder-dev",
-    "start:dev": "webpack-dev-server --hot --progress --color --mode development --config ../../scripts/webpack/webpack.sass.dev.js",
+    "start:dev": "webpack-dev-server --hot --progress --color --mode development --config ../../scripts/webpack/webpack.config.dev.js",
     "start:prod": "serve public/dist -s -l 8080",
     "test:ci": "cross-env TZ=UTC jest --ci --colors --coverage --silent --forceExit --detectOpenHandles --runInBand",
     "test:dev": "cross-env TZ=UTC jest --watch --verbose",
