+++ packages/smb-onboarder/package.json
@@ -11,8 +11,8 @@
   "author": "reapit global",
   "main": "dist/index.ts",
   "scripts": {
-    "build:prod": "rimraf public/dist && node ../../scripts/webpack/webpack-sass-prod/index.js",
-    "start:dev": "webpack-dev-server --hot --progress --color --mode development --config ../../scripts/webpack/webpack.sass.dev.js",
+    "build:prod": "rimraf public/dist && webpack --config ../../scripts/webpack/webpack.config.prod.js",
+    "start:dev": "webpack-dev-server --hot --progress --color --mode development --config ../../scripts/webpack/webpack.config.dev.js",
     "start:prod": "serve public/dist -s -l 8080",
     "test:ci": "cross-env TZ=UTC jest --ci --colors --coverage --silent --forceExit",
     "test:dev": "cross-env TZ=UTC jest --watch --verbose",
