--- scripts/webpack/sw.js
@@ -1,15 +1,0 @@
-// https://github.com/NekR/self-destroying-sw
-self.addEventListener('install', function() {
-  self.skipWaiting()
-})
-
-self.addEventListener('activate', function() {
-  self.registration
-    .unregister()
-    .then(function() {
-      return self.clients.matchAll()
-    })
-    .then(function(clients) {
-      clients.forEach(client => client.navigate(client.url))
-    })
-})
